拆分终端会话控制器
